+++ R/multistep_mars.R
@@ -450,6 +450,13 @@ predict.mars_multistep_fit_impl <- function(object, new_data, ...) {
 #' @export
 mars_multistep_predict_impl <- function(object, new_data, ...) {
 
+  # Date Mapping Table
+  date_tbl <- new_data %>%
+    dplyr::select(Date, Date_index.num) %>%
+    dplyr::distinct() %>%
+    dplyr::arrange(Date) %>%
+    dplyr::mutate(Run_Number = dplyr::row_number())
+
   # PREPARE INPUTS
   xreg_recipe <- object$extras$xreg_recipe
   h_horizon <- nrow(new_data)
@@ -458,14 +465,16 @@ mars_multistep_predict_impl <- function(object, new_data, ...) {
   xreg_tbl <- modeltime::bake_xreg_recipe(xreg_recipe,
     new_data,
     format = "tbl"
-  )
+  ) %>%
+    dplyr::left_join(date_tbl, by = "Date_index.num") %>%
+    dplyr::mutate(Row_Num = dplyr::row_number())
 
   # PREDICTIONS
   final_prediction <- tibble::tibble()
   start_val <- 1
 
   for (model_name in names(object$models)) {
-    if (start_val > nrow(xreg_tbl)) {
+    if (start_val > nrow(date_tbl)) {
       break
     }
 
@@ -474,7 +483,10 @@ mars_multistep_predict_impl <- function(object, new_data, ...) {
     mars_model <- object$models[[model_name]]
 
     xreg_tbl_final <- xreg_tbl %>%
-      dplyr::slice(start_val:lag_number)
+      dplyr::filter(
+        Run_Number >= start_val,
+        Run_Number <= lag_number
+      )
 
     if (!is.null(xreg_tbl)) {
       preds_mars <- predict(mars_model, xreg_tbl_final)
@@ -482,9 +494,17 @@ mars_multistep_predict_impl <- function(object, new_data, ...) {
       preds_mars <- rep(0, h_horizon)
     }
 
+    preds_mars <- preds_mars %>%
+      dplyr::mutate(Row_Num = xreg_tbl_final$Row_Num)
+
     start_val <- as.numeric(lag_number) + 1
     final_prediction <- rbind(final_prediction, preds_mars)
   }
+
+  # Ensure it's sorted correctly for global models
+  final_prediction <- final_prediction %>%
+    dplyr::arrange(Row_Num) %>%
+    dplyr::select(.pred)
 
   return(final_prediction)
 }